requirements.txt
flask==3.0.0
line-bot-sdk==3.21.0
firebase-admin==6.5.0
python-dotenv==1.0.0
